refactor: add back in toolbar components
