+++ components/ui/DesktopLabel.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const DesktopLabel = styled.span`
+  display: none;
+  @media (min-width: 1030px) {
+    display: inline;
+  }
+`
