+++ components/ui/ToolbarButton.tsx
@@ -1,0 +1,28 @@
+import styled from 'styled-components'
+import { Button } from '@tinacms/styles'
+
+export const ToolbarButton = styled(Button)`
+  display: flex;
+  align-items: center;
+  white-space: nowrap;
+  padding: 0 10px;
+  &:focus {
+    outline: none;
+  }
+  svg {
+    fill: currentColor;
+    opacity: 0.7;
+    width: 2.5em;
+    height: 2.5em;
+  }
+  &:disabled {
+    opacity: 0.6;
+    filter: grayscale(25%);
+  }
+  @media (min-width: 1030px) {
+    padding: 0 20px;
+    svg {
+      margin-right: 0.25rem;
+    }
+  }
+`
