+++ open-authoring/useOpenAuthoringToolbarPlugins.tsx
@@ -1,7 +1,7 @@
 import { Form, useCMS, FieldMeta } from 'tinacms'
 import { useEffect, useState } from 'react'
-import { ToolbarButton } from '../components/ui/inline/ToolbarButton'
-import { DesktopLabel } from '../components/ui/inline/DesktopLabel'
+import { ToolbarButton } from '../components/ui/ToolbarButton'
+import { DesktopLabel } from '../components/ui/DesktopLabel'
 import { PRPlugin } from './PRPlugin'
 import styled, { css } from 'styled-components'
 import { color } from '@tinacms/styles'
